social_media_api/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-environ==0.12.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
jsonfield==3.2.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
whitenoise==6.11.0
